+++ modules/Users/TeacherPrograms.php
@@ -20,29 +20,31 @@ if (UserStaffID())
 	echo '<form action="Modules.php?modname='.$_REQUEST['modname'].'" method="POST">';
 	//FJ multiple school periods for a course period
 	//$QI = DBQuery("SELECT cp.PERIOD_ID,cp.COURSE_PERIOD_ID,sp.TITLE,sp.SHORT_NAME,cp.MARKING_PERIOD_ID,cp.DAYS,c.TITLE AS COURSE_TITLE FROM COURSE_PERIODS cp,SCHOOL_PERIODS sp,COURSES c WHERE c.COURSE_ID=cp.COURSE_ID AND cp.PERIOD_ID=sp.PERIOD_ID AND cp.SYEAR='".UserSyear()."' AND cp.SCHOOL_ID='".UserSchool()."' AND cp.TEACHER_ID='".UserStaffID()."' AND cp.MARKING_PERIOD_ID IN (".GetAllMP('QTR',UserMP()).") ORDER BY sp.SORT_ORDER");
-	$QI = DBQuery("SELECT cpsp.PERIOD_ID,cp.COURSE_PERIOD_ID,cpsp.COURSE_PERIOD_SCHOOL_PERIODS_ID,sp.TITLE,sp.SHORT_NAME,cp.MARKING_PERIOD_ID,cpsp.DAYS,c.TITLE AS COURSE_TITLE, cp.SHORT_NAME AS CP_SHORT_NAME 
-	FROM COURSE_PERIODS cp,SCHOOL_PERIODS sp,COURSES c,COURSE_PERIOD_SCHOOL_PERIODS cpsp 
-	WHERE cp.COURSE_PERIOD_ID=cpsp.COURSE_PERIOD_ID 
-	AND c.COURSE_ID=cp.COURSE_ID 
-	AND cpsp.PERIOD_ID=sp.PERIOD_ID 
-	AND cp.SYEAR='".UserSyear()."' 
-	AND cp.SCHOOL_ID='".UserSchool()."' 
-	AND cp.TEACHER_ID='".UserStaffID()."' 
-	AND cp.MARKING_PERIOD_ID IN (".GetAllMP('QTR',UserMP()).") 
+	$QI = DBQuery("SELECT cpsp.PERIOD_ID,cp.COURSE_PERIOD_ID,cpsp.COURSE_PERIOD_SCHOOL_PERIODS_ID,sp.TITLE,sp.SHORT_NAME,cp.MARKING_PERIOD_ID,cpsp.DAYS,c.TITLE AS COURSE_TITLE, cp.SHORT_NAME AS CP_SHORT_NAME
+	FROM COURSE_PERIODS cp,SCHOOL_PERIODS sp,COURSES c,COURSE_PERIOD_SCHOOL_PERIODS cpsp
+	WHERE cp.COURSE_PERIOD_ID=cpsp.COURSE_PERIOD_ID
+	AND c.COURSE_ID=cp.COURSE_ID
+	AND cpsp.PERIOD_ID=sp.PERIOD_ID
+	AND cp.SYEAR='".UserSyear()."'
+	AND cp.SCHOOL_ID='".UserSchool()."'
+	AND cp.TEACHER_ID='".UserStaffID()."'
+	AND cp.MARKING_PERIOD_ID IN (".GetAllMP('QTR',UserMP()).")
 	ORDER BY cp.SHORT_NAME, sp.SORT_ORDER");
-	
+
 	$RET = DBGet($QI);
 	// get the fy marking period id, there should be exactly one fy marking period
 	$fy_RET = DBGet(DBQuery("SELECT MARKING_PERIOD_ID FROM SCHOOL_MARKING_PERIODS WHERE MP='FY' AND SYEAR='".UserSyear()."' AND SCHOOL_ID='".UserSchool()."'"));
 
-	if ( $_REQUEST['period'])
+	if ( $_REQUEST['period'] )
 	{
-		list($CoursePeriod, $CoursePeriodSchoolPeriod) = explode('.', $_REQUEST['period']);
+		list( $CoursePeriod, $CoursePeriodSchoolPeriod ) = explode( '.', $_REQUEST['period'] );
+
 		$_SESSION['UserCoursePeriod'] = $CoursePeriod;
+
 		$_SESSION['UserCoursePeriodSchoolPeriod'] = $CoursePeriodSchoolPeriod;
 	}
 
-	if ( !UserCoursePeriod())
+	if ( ! UserCoursePeriod() )
 	{
 		$_SESSION['UserCoursePeriod'] = $RET[1]['COURSE_PERIOD_ID'];
 		$_SESSION['UserCoursePeriodSchoolPeriod'] = $RET[1]['COURSE_PERIOD_SCHOOL_PERIODS_ID'];
@@ -71,7 +73,7 @@ if (UserStaffID())
 		else
 			$selected = '';
 
-		//FJ days display to locale						
+		//FJ days display to locale
 		$days_convert = array('U' => _('Sunday'),'M' => _('Monday'),'T' => _('Tuesday'),'W' => _('Wednesday'),'H' => _('Thursday'),'F' => _('Friday'),'S' => _('Saturday'));
 		//FJ days numbered
 		if (SchoolInfo('NUMBER_DAYS_ROTATION') !== null)
@@ -82,7 +84,7 @@ if (UserStaffID())
 		for ($i = 0; $i < $days_strlen; $i++) {
 			$period_DAYS_locale .= mb_substr($days_convert[mb_substr($period['DAYS'], $i, 1)],0,3) . '.';
 		}
-		
+
 		//FJ add subject areas
 		//$period_select .= '<option value="'.$period['COURSE_PERIOD_ID'].'.'.$period['COURSE_PERIOD_SCHOOL_PERIODS_ID'].'"'.$selected.'>'.$period['SHORT_NAME'].(mb_strlen($period['DAYS'])<5?' ('.$period_DAYS_locale.')':'').($period['MARKING_PERIOD_ID']!=$fy_RET[1]['MARKING_PERIOD_ID']?' '.GetMP($period['MARKING_PERIOD_ID'],'SHORT_NAME'):'').' - '.$period['CP_SHORT_NAME'].'</option>';
 		$period_select .= '<option value="'.$period['COURSE_PERIOD_ID'].'.'.$period['COURSE_PERIOD_SCHOOL_PERIODS_ID'].'"'.$selected.'>'.$period['TITLE'].(mb_strlen($period['DAYS'])<5?(mb_strlen($period['DAYS'])<2?' '._('Day').' '.$period_DAYS_locale.' - ':' '._('Days').' '.$period_DAYS_locale.' - '):' - ').($period['MARKING_PERIOD_ID']!=$fy_RET[1]['MARKING_PERIOD_ID']?GetMP($period['MARKING_PERIOD_ID'],'SHORT_NAME').' - ':'').$period['CP_SHORT_NAME'].'</option>';